fix(propostas): cast para ::numeric antes do ROUND para compatibilidade PostgreSQL
ROUND(float8, int) não existe no PostgreSQL — apenas ROUND(numeric, int).
A query retornava erro 500 em toda chamada, zerando o grid do dashboard.

+++ backend/handlers/sp_propostas.go
@@ -115,13 +115,13 @@ func SpPropostasHandler(db *sql.DB) http.HandlerFunc {
 			       TO_CHAR(p.created_at,'YYYY-MM-DD"T"HH24:MI:SS"Z"'),
 			       CASE
 			         WHEN e.qt_acesso_90 > 0 AND e.qt_dias > 0
-			           THEN ROUND(e.qt_acesso_90::float8 / NULLIF(e.qt_dias,0), 3)
+			           THEN ROUND((e.qt_acesso_90::float8 / NULLIF(e.qt_dias,0))::numeric, 3)
 			         WHEN e.med_venda_dias > 0
-			           THEN ROUND(e.med_venda_dias::float8, 3)
+			           THEN ROUND(e.med_venda_dias::numeric, 3)
 			         WHEN e.med_venda_cx > 0 AND e.unidade_master > 0
-			           THEN ROUND(e.med_venda_cx * e.unidade_master, 3)
+			           THEN ROUND((e.med_venda_cx * e.unidade_master)::numeric, 3)
 			         WHEN e.med_venda_cx_aa > 0 AND e.unidade_master > 0
-			           THEN ROUND(e.med_venda_cx_aa * e.unidade_master, 3)
+			           THEN ROUND((e.med_venda_cx_aa * e.unidade_master)::numeric, 3)
 			         ELSE NULL
 			       END,
 			       e.med_venda_cx,
